Move LVM creation from LvmHelper to its own class
