+++ src/lib/y2storage/proposal.rb
@@ -31,6 +31,7 @@ require "y2storage/proposal/autoinst_devices_planner"
 require "y2storage/proposal/autoinst_space_maker"
 require "y2storage/proposal/autoinst_drives_map"
 require "y2storage/proposal/partition_creator"
+require "y2storage/proposal/lvm_creator"
 require "y2storage/proposal/partition_killer"
 require "y2storage/proposal/partitions_distribution_calculator"
 require "y2storage/proposal/phys_vol_calculator"
